requirements.txt
asgiref==3.2.7
boto3==1.13.1
botocore==1.16.1
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
psycopg2==2.8.2
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1

